[IMP] im_livechat: add kanban view for session history
Purpose of this commit:
Introduce the Kanban view for live chat session history to enable
easy navigation through sessions.

task-4043438
