+++ addons/im_livechat/static/src/views/discuss_channel_list/discuss_channel_list_view.js
@@ -1,7 +1,9 @@
-import { DiscussChannelListController } from "@im_livechat/views/discuss_channel_list/discuss_channel_list_view_controller";
-
+import { ListController } from "@web/views/list/list_controller";
+import { LivechatViewControllerMixin } from "../livechat_view_controller_mixin";
 import { listView } from "@web/views/list/list_view";
 import { registry } from "@web/core/registry";
+
+class DiscussChannelListController extends LivechatViewControllerMixin(ListController) {}
 
 const discussChannelListView = {
     ...listView,
